- Создание необходимых директорий и расфасовка файлов - Рефакторинг визуальной части приложения
