--- Stars_Game/MyFirstWPFapp/MainWindow.xaml
@@ -1,58 +1,0 @@
-﻿<Window x:Class="MyFirstWPFapp.MainWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:sys="clr-namespace:System;assembly=mscorlib"
-        xmlns:local="clr-namespace:MyFirstWPFapp"
-        mc:Ignorable="d"
-        Title="Департаменты и сотрудники" Height="503.369" Width="800"
-        Background="AntiqueWhite">
-
-    <Window.Resources>
-
-        <LinearGradientBrush x:Key="WindowBackgroundBrush">
-            <GradientStop Offset="0" Color="Silver"/>
-            <GradientStop Offset="1" Color="AliceBlue"/>
-        </LinearGradientBrush>
-    </Window.Resources>
-
-    <StackPanel Margin="10">
-        <StackPanel.Resources>
-            <sys:String x:Key="ListBoxTitle">Список сотрудников</sys:String>
-            <sys:String x:Key="ListBoxTitle1">Список департаментов</sys:String>
-        </StackPanel.Resources>
-
-        <Label Content="{StaticResource ListBoxTitle1}"/>
-        <Label Content="{StaticResource ListBoxTitle}"/>
-
-        <Grid>
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition Width="9*"/>
-                <ColumnDefinition Width="0*"/>
-                <ColumnDefinition Width="13*"/>
-                <ColumnDefinition Width="8*"/>
-                <ColumnDefinition Width="5*"/>
-                <ColumnDefinition Width="9*"/>
-                <ColumnDefinition Width="51*"/>
-                <ColumnDefinition Width="0*"/>
-                <ColumnDefinition Width="0*"/>
-                <ColumnDefinition Width="291*"/>
-            </Grid.ColumnDefinitions>
-            <ListBox Margin="10,20,10,0" Name="lbEmployee" SelectionMode="Single"
-                SelectionChanged="lbEmployee_SelectionChanged"
-             VerticalAlignment="Top" Height="250" Grid.ColumnSpan="10"></ListBox>
-            <Button Margin="18,0,430,-78" Height="28" VerticalAlignment="Bottom"
-                    Click="Button_Click" Grid.Column="2" Grid.ColumnSpan="8">Новый сотрудник</Button>
-
-            <ComboBox Name="lbDepartament"
-                      VerticalAlignment="Top" Height="22" Margin="65,-50,430,0" Grid.ColumnSpan="4" Grid.Column="6"/>
-            <Button Margin="218,0,37,-78" Height="28" VerticalAlignment="Bottom"
-                    Click="Button_Click_dep" Grid.Column="9">Новый департамент</Button>
-            <Button Margin="46,0,261,-121" Height="22" VerticalAlignment="Bottom"
-Click="Button_Click_1" Grid.Column="9">Выход</Button>
-        </Grid>
-    </StackPanel>
-
-
-</Window>
