+++ Stars_Game/MyFirstWPFapp/Views/MainWindow.xaml
@@ -1,0 +1,46 @@
+﻿<Window x:Class="MyFirstWPFapp.MainWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:sys="clr-namespace:System;assembly=mscorlib"
+        xmlns:local="clr-namespace:MyFirstWPFapp"
+        mc:Ignorable="d"
+        Title="Департаменты и сотрудники" Height="503.369" Width="800"
+        Background="AntiqueWhite">
+
+    <Window.Resources>
+        <LinearGradientBrush x:Key="WindowBackgroundBrush">
+            <GradientStop Offset="0" Color="Silver"/>
+            <GradientStop Offset="1" Color="AliceBlue"/>
+        </LinearGradientBrush>
+        
+        
+    </Window.Resources>
+    <DockPanel>
+        <Menu DockPanel.Dock="Top">
+            <MenuItem Header="Файл">
+                <MenuItem Header="Закрыть"></MenuItem>
+            </MenuItem>
+            <MenuItem Header="О программе"></MenuItem>
+        </Menu>
+        <TabControl>
+            <TabItem Header="Департаменты">
+                <Grid>
+                    <Grid.ColumnDefinitions>
+                        <ColumnDefinition Width="3*"/>
+                        <ColumnDefinition Width="7*"/>
+                    </Grid.ColumnDefinitions>
+                    <GroupBox Header="Департамент">
+                        <ListBox></ListBox>
+                    </GroupBox>
+                    <GroupBox Header="Сотрудники департамента" Grid.Column="1">
+                        <ListBox></ListBox>
+                    </GroupBox>
+                </Grid>
+            </TabItem>
+            <TabItem Header="Сотрудники"></TabItem>
+        </TabControl>
+    </DockPanel>
+
+</Window>
